Fixing bug for CUDA tensors

+++ helperLayers.py
@@ -3,7 +3,7 @@ import torchsummary
 import numpy as np
 from torch import nn
 
-torch.set_default_tensor_type(torch.cuda.DoubleTensor)
+torch.set_default_tensor_type(torch.cuda.FloatTensor)
 
 class ResBlock(nn.Module):
     def __init__(self,kernel_size,channels,padding,stride):
