+++ loss.py
@@ -4,7 +4,7 @@ import numpy as np
 from torch import nn
 from torchvision.models import vgg19
 
-torch.set_default_tensor_type(torch.cuda.DoubleTensor)
+torch.set_default_tensor_type(torch.cuda.FloatTensor)
 
 class PerceptualLoss(nn.Module):
     def __init__(self):
