+++ model.py
@@ -4,7 +4,7 @@ import numpy as np
 from torch import nn
 from helperLayers import *
 
-torch.set_default_tensor_type(torch.cuda.DoubleTensor)
+torch.set_default_tensor_type(torch.cuda.FloatTensor)
 
 class Generator(nn.Module):
     def __init__(self):
